# engageiu/openapi.yaml
openapi: "3.0.3"
info:
  title: EngageIU API
  version: "1.0.0"
  description: >
    QR/code-based campus event attendance leaderboard system for Indiana University.
    Students earn points by entering check-in codes at IU events.
    Built for Luddy Hacks — Dynamic Leaderboard/Ranking System case.
  contact:
    name: EngageIU Team
    email: iuhacks@iu.edu
servers:
  - url: http://localhost:8000
    description: Local development server
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Hoosier Innovation Summit"
        description:
          type: string
          example: "Annual innovation event for IU students."
        campus:
          type: string
          example: "IU Bloomington"
        event_url:
          type: string
          example: "https://events.iu.edu"
        points:
          type: integer
          example: 20
        event_date:
          type: string
          format: date-time
          example: "2025-04-14T10:00:00Z"
        attendees:
          type: integer
          example: 42
    EventWithCode:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            check_in_code:
              type: string
              example: "INNOV2025"
...
paths:
...
  /events:
    get:
      tags: [events]
      summary: List events
      description: >
        Public: returns event list without check_in_code.
        Admin: includes check_in_code for each event.
      parameters:
        - in: query
          name: campus
          schema:
            type: string
        - in: query
          name: upcoming_only
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

    post:
      tags: [events]
      summary: Create event (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, campus, event_date]
              properties:
                title:
                  type: string
                  example: "IU Career Fair"
                description:
                  type: string
                campus:
                  type: string
                  example: "IU Bloomington"
                event_url:
                  type: string
                points:
                  type: integer
                  default: 10
                event_date:
                  type: string
                  format: date-time
                  example: "2025-04-20T10:00:00Z"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithCode'
        "401":
          description: Unauthorized
  /events/{event_id}:
    patch:
      tags: [events]
      summary: Update event (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                campus:
                  type: string
                event_url:
                  type: string
                points:
                  type: integer
                event_date:
                  type: string
                  format: date-time
                regenerate_code:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithCode'
        "401":
          description: Unauthorized
        "404":
          description: Event not found

    delete:
      tags: [events]
      summary: Delete event (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
        "401":
          description: Unauthorized
        "404":
          description: Event not found
  /events/{event_id}/code:
    get:
      tags: [events]
      summary: Get check-in code for event (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in code
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: integer
                  check_in_code:
                    type: string
                    example: "INNOV2025"
        "401":
          description: Unauthorized
        "404":
          description: Event not found

  # ── STUDENT CHECK-IN ──────────────────────────────────────────────────────
  /checkin:
    post:
      tags: [events]
      summary: Student self check-in
      description: >
        Public endpoint. Student provides their info and the event check-in code.
        Creates a student record if this is their first check-in.
        Returns their updated weekly rank and total points.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [iu_username, name, campus, check_in_code]
              properties:
                iu_username:
                  type: string
                  example: "jdoe"
                name:
                  type: string
                  example: "Jane Doe"
                campus:
                  type: string
                  example: "IU Bloomington"
                check_in_code:
                  type: string
                  example: "INNOV2025"
      responses:
        "200":
          description: Check-in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  points_earned:
                    type: integer
                    example: 20
                  total_points:
                    type: integer
                    example: 35
                  current_rank:
                    type: integer
                    example: 3
                  event:
                    type: object
                    properties:
                      title:
                        type: string
                      campus:
                        type: string
        "404":
          description: Invalid check-in code
        "409":
          description: Already checked in to this event

  # ── STUDENTS ──────────────────────────────────────────────────────────────
  /students/search:
    get:
      tags: [students]
      summary: Search students
      description: Search by partial name or IU username. Returns weekly points and rank.
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Partial name or IU username
        - in: query
          name: campus
          schema:
            type: string
      responses:
        "200":
          description: Matching students
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    iu_username:
                      type: string
                    campus:
                      type: string
                    weekly_points:
                      type: integer